pyproject.toml
[tool.poetry]
name = "insta-video-gen"
version = "0.1.0"
description = "Instragram Video Generator"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "insta_video_gen"}]

[tool.poetry.dependencies]
python = "^3.11"
instaloader = "^4.10"
moviepy = "^1.0.3"
pydantic-settings = "^2.0.2"
pyright = "^1.1.318"
python-dateutil = "^2.8.2"
pillow = "9.5.0"
selenium = "^4.11.2"
google-auth = "^2.22.0"
google-auth-oauthlib = "^1.0.0"
google-auth-httplib2 = "^0.1.0"
google-api-python-client = "^2.97.0"
pydantic = "^2.3.0"
click = "^8.1.7"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
